refactor: move generic testing utilities into utils/testing from release/testing
Many of the util functions and classes from release/testing could be leveraged elsewhere in the
project and would be better served being located in utils/testing.
