+++ ng-dev/release/publish/test/test-utils/test-utils.ts
@@ -12,9 +12,9 @@ import * as semver from 'semver';
 import {getTestConfigurationsForAction, setupMocksForReleaseAction} from './action-mocks';
 import {_npmPackageInfoCache, ActiveReleaseTrains, NpmPackageInfo} from '../../../versioning';
 import {ReleaseAction, ReleaseActionConstructor} from '../../actions';
-import {GithubTestingRepo} from './github-api-testing';
+import {GithubTestingRepo} from '../../../../utils/testing';
 import {defaultTestOptions, TestOptions, TestReleaseAction} from './test-action';
-import {dedent} from '../../../../utils/testing/dedent';
+import {dedent} from '../../../../utils/testing';
 import {testTmpDir} from '../../../../utils/testing';
 
 /**
