+++ ng-dev/utils/testing/BUILD.bazel
@@ -12,6 +12,7 @@ ts_library(
     deps = [
         ":bazel-env",
         "//ng-dev/utils",
+        "@npm//@octokit/plugin-rest-endpoint-methods",
         "@npm//@types/jasmine",
         "@npm//@types/node",
         "@npm//@types/semver",
