+++ ng-dev/utils/testing/index.ts
@@ -10,3 +10,9 @@ export * from './semver-matchers';
 export * from './virtual-git-client';
 export * from './virtual-git-matchers';
 export * from './bazel-env';
+export * from './sandbox-git-client';
+export * from './sandbox-testing';
+export * from './git-client-mock';
+export * from './github-api-testing';
+export * from './github-interception';
+export * from './dedent';
